Added ceiling to tau calculation

+++ exp.h
@@ -3,6 +3,6 @@
 
 void set_exp_params(double lambda, int threshold, double alpha);
 double next_exp();
-double next_tau(double prev_tau, double tau);
+int next_tau(int prev_tau, int tau);
 
 #endif
